requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
et-xmlfile==1.1.0
fastapi==0.101.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
mysql-connector==2.2.9
numpy==1.25.2
openai==0.27.8
openpyxl==3.1.2
pandas==2.0.3
psycopg2==2.9.7
pydantic==2.1.1
pydantic_core==2.4.0
pyodbc==4.0.39
pyscopg2==66.0.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rivescript==1.15.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2